Revisions to two project entries
I updated the copy for Seattle and PPH. I could not go forward because
there are is no structure in place in yml for the rest of the projects.
I considered cutting and pasting one of the existing entries to serve
as a model for entering the others, but each entry needs an “anchor”
and I did not know what that would be.

I’ll tackle this again on Monday.