Update data.yml
2nd round of updates

Added mention of CNCF graduation in Projects

Removed the word "recent" from description of Uptane award.

Added reference to JDF under Uptane

Cleaned up a few links in the People section

Replace references to Uptane Google Doc and added the Standard to the Publications page

Added 17 press references for TUF, Uptane, and in-toto, bringing that section up to date.